requirements.txt
bypy==1.5.7
certifi==2017.4.17
chardet==3.0.4
dill==0.2.6
Django==1.10.5
django-crontab==0.7.1
django-guardian==1.4.9
idna==2.5
multiprocess==0.70.5
olefile==0.44
pexpect==4.3.0
Pillow==4.2.0
psutil==5.2.2
ptyprocess==0.5.2
PyMySQL==0.7.11
pytz==2017.2
requests==2.18.1
requests-toolbelt==0.8.0
scapy-python3==0.21
six==1.10.0
tzlocal==1.4
urllib3==1.21.1
